+++ edu.kit.ipd.are.smarthomedata/src/main/java/edu/kit/ipd/are/smarthomedata/old/OutliersAllHouses.java
@@ -17,15 +17,15 @@ import edu.kit.ipd.are.smarthomedata.dto.SmartMeterReading;
  *
  */
 public class OutliersAllHouses implements Consumer<SmartMeterReading> {
-    private final Map<Long, Float> houses;
+    private final Map<Long, Double> houses;
 	private Consumer<Double> callback;
 
-    public float outliers() {
-        final Collection<Float> values = houses.values();
-        Float[] medians = new Float[values.size()];
+    public double outliers() {
+        final Collection<Double> values = houses.values();
+        Double[] medians = new Double[values.size()];
         Arrays.parallelSort(values.toArray(medians));
 
-        float median = 0.0F;
+        double median = 0.0F;
         if ((medians.length % 2) == 0) {
             median = (medians[medians.length / 2] + medians[(medians.length / 2) - 1]) / 2;
         } else {
@@ -39,7 +39,7 @@ public class OutliersAllHouses implements Consumer<SmartMeterReading> {
             }
         }
 
-        return (float) count / medians.length;
+        return (double) count / medians.length;
     }
 
     @Override
@@ -47,7 +47,7 @@ public class OutliersAllHouses implements Consumer<SmartMeterReading> {
         houses.put(t.house_id, t.value);
     }
 
-    public OutliersAllHouses(Map<Long, Float> houses, Consumer<Double> callback) {
+    public OutliersAllHouses(Map<Long, Double> houses, Consumer<Double> callback) {
         this.houses = houses;
         this.callback = callback;
     }
